requirements.txt
aiogram==2.25.1
python-dotenv==1.0.1
loguru==0.6.0
peewee~=3.15.4
peewee-migrate~=1.6.6
